Keep OpenTelemetry SDK dependencies provider-local
